requirements.txt
altgraph==0.17.2
future==0.18.2
paho-mqtt==1.6.1
pefile==2022.5.30
pyinstaller>=4.10
pyinstaller-hooks-contrib==2022.0
pywin32-ctypes==0.2.0
